+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/models/CompletionsUsage.java
@@ -91,6 +91,8 @@ public final class CompletionsUsage implements JsonSerializable<CompletionsUsage
         jsonWriter.writeIntField("completion_tokens", this.completionTokens);
         jsonWriter.writeIntField("prompt_tokens", this.promptTokens);
         jsonWriter.writeIntField("total_tokens", this.totalTokens);
+        jsonWriter.writeJsonField("prompt_tokens_details", this.promptTokensDetails);
+        jsonWriter.writeJsonField("completion_tokens_details", this.completionTokensDetails);
         return jsonWriter.writeEndObject();
     }
 
@@ -109,6 +111,8 @@ public final class CompletionsUsage implements JsonSerializable<CompletionsUsage
             int completionTokens = 0;
             int promptTokens = 0;
             int totalTokens = 0;
+            CompletionsUsagePromptTokensDetails promptTokensDetails = null;
+            CompletionsUsageCompletionTokensDetails completionTokensDetails = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
@@ -118,11 +122,51 @@ public final class CompletionsUsage implements JsonSerializable<CompletionsUsage
                     promptTokens = reader.getInt();
                 } else if ("total_tokens".equals(fieldName)) {
                     totalTokens = reader.getInt();
+                } else if ("prompt_tokens_details".equals(fieldName)) {
+                    promptTokensDetails = CompletionsUsagePromptTokensDetails.fromJson(reader);
+                } else if ("completion_tokens_details".equals(fieldName)) {
+                    completionTokensDetails = CompletionsUsageCompletionTokensDetails.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
             }
-            return new CompletionsUsage(completionTokens, promptTokens, totalTokens);
+            CompletionsUsage deserializedCompletionsUsage
+                = new CompletionsUsage(completionTokens, promptTokens, totalTokens);
+            deserializedCompletionsUsage.promptTokensDetails = promptTokensDetails;
+            deserializedCompletionsUsage.completionTokensDetails = completionTokensDetails;
+            return deserializedCompletionsUsage;
         });
+    }
+
+    /*
+     * Details of the prompt tokens.
+     */
+    @Generated
+    private CompletionsUsagePromptTokensDetails promptTokensDetails;
+
+    /*
+     * Breakdown of tokens used in a completion.
+     */
+    @Generated
+    private CompletionsUsageCompletionTokensDetails completionTokensDetails;
+
+    /**
+     * Get the promptTokensDetails property: Details of the prompt tokens.
+     *
+     * @return the promptTokensDetails value.
+     */
+    @Generated
+    public CompletionsUsagePromptTokensDetails getPromptTokensDetails() {
+        return this.promptTokensDetails;
+    }
+
+    /**
+     * Get the completionTokensDetails property: Breakdown of tokens used in a completion.
+     *
+     * @return the completionTokensDetails value.
+     */
+    @Generated
+    public CompletionsUsageCompletionTokensDetails getCompletionTokensDetails() {
+        return this.completionTokensDetails;
     }
 }